feat: Create and use own get_client_ip (for rate-limiting & more)
Instead of using a default of XFF[0] or XFF[-1] or REMOTE_ADDR[0], all of
which are certainly wrong for someone, make IP determination customizable
(but fall back to a good default of XFF-walking.)

This removes direct use of the django-ipware package in edxapp. We weren't
using that package's XFF-chain-reading capabilities, and in any case it
isn't capable of handling trusted-proxy IP ranges in CIDR or other formats
that would be required for handling e.g. Cloudflare IPs (instead requiring
textual prefixes). It also has a handful of bugs that make it difficult to
use safely.

Main changes:

- New `get_client_ip` that first tries to use specified headers from new
  `CLIENT_IP_HEADERS` setting, then falls back to walking XFF.
- Stop using django-ipware, in favor of ipaddress module and new code
- Save off original `REMOTE_ADDR` before overwriting it
- Add some comments warning people away from relying on REMOTE_ADDR
  directly until we can remove that middleware

Related changes:

- Fix test that had wrong idea of XFF

Also:

- Allow some safe-looking isort auto-fixes
- Some unused code cleanup (remaining headers don't contain commas) and
  delinting.

Internal ticket: ARCHBOM-2025